applet: add support for --with-appindicator
Add new configure option --with-appindicator which implies a optional
dependency on libappindicator. This is the way systray is done in the
current versions of Unity, KDE and Enlightenment and others. Some of the
systray implementations already dropped support for Xembed, others are
going to, mostly to gain independence on X11 as part of being ported to
Wayland.

Keep the previous functionality intact.